fixup! The EVP_MAC functions have been renamed for consistency.  The EVP_MAC_CTX_* functions are now EVP_MAC functions, usually with ctx in their names.
